+++ NEWS.md
@@ -1,3 +1,10 @@
+# conserver 1.0.2
+
+* added new methods and updated literature and traits data
+* added `map_selection` function
+* added mca_results and connectivity_network data
+* included scalability measure into traits data
+
 # conserver 1.0.1
 
 * update literature dataset